Edited 'main.py'

+++ telegraf-normalisation/main.py
@@ -37,7 +37,7 @@ def convert_to_sensor_table(row: dict):
         yield output_row
 
 sdf = sdf.apply(convert_to_sensor_table, expand=True)
-sdf = sdf.set_timestamp(lambda row, *_: row["timestamp"] / 1E6)
+sdf = sdf.set_timestamp(lambda row, *_: int(row["timestamp"] / 1E6))
 sdf.print()
 
 sdf.to_topic(output_topic)
